requirements.txt
google-generativeai
python-dotenv
pyyaml
python-frontmatter
jinja2
PyQt6
